Reduce AddItem and EditItem pages into a single class
